Move JobStatus class in its own file
